% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cons.test.R
\name{cons.test}
\alias{cons.test}
\title{Computes short cut for testing intersection hypotheses of m hypotheses}
\usage{
cons.test(p, method = "stouffer", Gnull, alp, func = NULL)
}
\arguments{
\item{p}{vector of p-values}

\item{method}{either "Stouffer" or "Fisher" combination test}

\item{Gnull}{list of null distribution, generated by create.null}

\item{alp}{significance level}

\item{func}{userdefined function. Is set to zero if no user defined function is given}
}
\value{
adjusted p-values
}
\description{
Computes short cut for testing intersection hypotheses of m hypotheses
}
