% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.null.R
\name{create.null}
\alias{create.null}
\title{Generation of null distribution of combination function to compute empirical p-values}
\usage{
create.null(
  method = "stouffer",
  alp,
  h,
  runs = 10^5,
  eps = 10^(-10),
  func = NULL
)
}
\arguments{
\item{method}{either "Stouffer" or "Fisher" combination test}

\item{alp}{significance level}

\item{h}{number of dimensions of null distribution}

\item{runs}{number of random variables under null hypothesis}

\item{eps}{auxiliary parameter}

\item{func}{userdefined function. Is set to zero if no user defined function is given}
}
\value{
list of null distributions
}
\description{
Generation of null distribution of combination function to compute empirical p-values
}
