% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fk.R
\name{fk}
\alias{fk}
\title{combination function resulting in (non-negative) test statistics}
\usage{
fk(p, method, func = NULL)
}
\arguments{
\item{p}{vector or matrix of p-values}

\item{method}{either "Stouffer" or "Fisher" combination test}

\item{func}{userdefined function. Is set to zero if no user defined function is given}
}
\value{
List of test statistics for each dimension
}
\description{
combination function resulting in (non-negative) test statistics
}
