% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pint.R
\name{pint}
\alias{pint}
\title{Computes a consonant test (p-value) for a specific intersection hypothesis iteratively}
\usage{
pint(p, method = "stouffer", Gnull, alp, func = NULL)
}
\arguments{
\item{p}{vector of p-values}

\item{method}{either "Stouffer" or "Fisher" combination test}

\item{Gnull}{list of null distribution, generated by create.null}

\item{alp}{significance level}

\item{func}{userdefined function. Is set to zero if no user defined function is given}
}
\value{
consonant p-value
}
\description{
Computes a consonant test (p-value) for a specific intersection hypothesis iteratively
}
